Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
yfinance==0.2.40
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.7.2
pydantic==2.8.2
PyPortfolioOpt==1.5.5
